pyproject.toml
[project]
name = "gemini-terminal-chat"
version = "0.1.0"
description = "A terminal-based chat interface for Google's Gemini AI"
requires-python = ">=3.12"
dependencies = [
  "google-genai>=1.2.0",
  "python-dotenv>=1.0.1",
  "colorama>=0.4.6",
  "rich>=13.9.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.pyright]
include = [".."]
extraPaths = [".."]
